cloudbrain/requirements.txt
cassandra-driver
blist
pika
flask